fermm-server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
pillow>=10.0.0
